`Development`: Merge test classes for same service
